pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polygon-qwen"
version = "0.1.0"
description = "Qwen3-VL-2B experiments"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "opencv-python>=4.9",
    "Pillow>=10.0",
    "accelerate>=1.0",
    "peft>=0.13",
    "torch>=2.3",
    "torchvision>=0.18",
    "transformers>=4.57.0",
]

[project.optional-dependencies]
npu = [
    "torch==2.7.1",
    "torch-npu==2.7.1",
    "torchvision==0.22.1",
    "transformers==4.57.1",
]

[tool.setuptools.packages.find]
where = ["src"]
